Projeto1_AutomacaoAnaliseDados/requirements.txt
pyautogui==0.9.54
pandas==2.2.3
plotly==5.24.1
nbformat==5.10.4
python-dotenv==1.0.1
